+++ cmake/nfxGraphicsSources.cmake
@@ -11,6 +11,7 @@ set(NFX_GRAPHICS_SOURCES
     src/gl/pipeline/passes/GeometryPass.cpp
     src/gl/pipeline/passes/GridPass.cpp
     src/gl/pipeline/passes/PresentPass.cpp
+    src/gl/pipeline/passes/SpotShadowPass.cpp
     src/gl/pipeline/passes/TransparentPass.cpp
     src/gl/pipeline/Renderer.cpp
     src/gl/pipeline/RenderState.cpp